feat: Create object candidates for sourcelink downloads
This integrates sourcelink downloads into the object
candidate machinery. Previously, if a download from
a sourcelink failed, there would be no indication that
anything had gone wrong. Now, whenever we try to
download a source file from a sourcelink, we create
a "sourcelink" candidate and add it to the module's
candidates. Since the object candidate system was not
set up for adding candidates after the fact, this
requires some tinkering.