feat(sentry): Add a config flag for propagating sentry traces (#1525)
The trace and span volume between Sentry and Symbolicator can differ
significantly, so in some scenarios trace propagation might not be
desirable. To give more control over the behavior, this PR introduces a
new config flag `propagate_traces`, that controls:

- Inheriting parent sample rates for incoming traces via HTTP. Traces
  will still be continued, but symbolicator's spans will be sampled at
  their own rate.
- Propagating outgoing traces to Sentry via the `sentry-trace` header

Additionally, this fixes the field name of the traces sample rate
config, as this was incorrectly called `transaction_sample_rate`. For
backwards compatibility, we still support an alias.