feat: Add `Unsupported` variant to `CacheError` and `ObjectFileStatus` (#1554)
This adds a new `CacheError`/`ObjectFileStatus`
variant for cases where a debug file is OK, but can't be used
for the purpose for which it was requested. Currently the only
such case is symbolicating .NET events with portable PDB files
(see getsentry/symbolic#871).

In addition, this also makes Symbolicator request both
portable PDB and Windows PDB files for symbolicating .NET
events. The reason for this is that although Windows PDB files
can never be used for that purpose, at least this way we get a
concrete error status we can report to the user. If we don't do
that, we get cases like #1539.