ref: Split apart native and JS modules
This is part of the effort to unify Symbolicator's
symbolication endpoints. It may seem strange to split
types apart in the pursuit of unification, so let
me explain. Currently, it's fine to use the same type
for native and JS modules because we are always using
them in the context of symbolicating exactly one of
these platforms. But once there is only one symbolication
endpoint, it will be convenient to be able to split
modules according to their platforms.