fix(symsorter): Don't create files twice
If, for some reason, there are two different debug files
with the same ID, we previously would open the target file
twice and write to it in parallel, which could cause data
corruption (only observed when using compression, which
might be a matter of speed). Now instead we only create
the file if it doesn't exist and print a warning otherwise.
The resulting warnings look like

> WARNING: File out-dir/bc/0f534e9ab5a4c7/debuginfo already exists, you seem to have duplicate debug files for ID bc0f534e9ab5a4c7.
> Skipping foobar.debug.