feat: Add metadata to cache entries (#1617)
This attaches an (optional) metadata struct to cache entries
produced by `compute_memoized`. For now, this structure contains two pieces of data:

1. The scope (project ID or "global") of the request.
2. The creation time of the entry.

We may add more in the future.

We don't currently write or read the metadata at any point.
When we create a CacheEntry we use the without_metadata
constructor, and when we consume one, we always use
contents() or into_contents() to get the data. Actually
using the metadata is left for a future PR.

This also renames the former CacheEntry to CacheContents
and uses the name CacheEntry for the new "cache contents + metadata" struct.