feat: Read and write metadata files (#1618)
This adds functionality to write and read the metadata structures introduced in #1617 to and from the filesystem. A cache file's metadata file has the same path, but with the extension `.metadata`.

The read happens in check_expiry, which gives the cleanup job access to the metadata. All of the old logic is still there—if no metadata file exists, we use the mtime from the FS as before.

The write happens in compute. It should be noted that currently metadata are not propagated through the shared cache; this is left as future work.