fix(nuget): Capture error when no debug checksum
The public NuGet symbol source returns a permission denied
error when trying to download a portable pdb file without
a debug checksum. We capture an error whenever we try to
download from this source without such a checksum to see
if this is the cause of our many permission denied errors.