ref: Log successful downloads from Electron (#1638)
Temporarily log whenever we successfully download
a file from the Electron symbol server. These are
few and far between, so the volume should be no
problem.

The purpose of this is to validate that we only
find files with very specific debug file names
(my current assumption). If that is true, we can
use path patterns to significantly limit what
we request from Electron.