feat(minidump): Introduce rewriting of debug file names for lookups (#1650)
# Rationale
In order to stackwalk and symbolicate minidumps from Electron, we have to
look up symbol files on the official Electron symbol server. The lookup of a module
involves the module's debug file name. Debug file names that can be found on
the Electron symbol server include "Electron Framework", "electron.exe.pdb",
and others.

The problem is that when an Electron application is bundled, the main executable
can be renamed, and this renaming affects minidumps. Specifically, the debug file name
of the first loaded module follows the executable name. For example, crashing a simple
test application may produce a minidump in which the first module has debug file "Electron".
But bundling the test app as "My Awesome Crasher" causes the first module's debug file to
also be called "My Awesome Crasher".

This renaming causes the debug file not to be found on the Electron source under the modified
name, even though it is available under the original name. This is obviously unsatisfactory.

# Solution
This PR adds a new optional parameter `rewrite_first_module` to the minidump
endpoint. This can be used to pass Symbolicator a list of _rewrite rules_
in the form of a regex and a replacement string. These rewrite rules will be used
during minidump stackwalking and symbolication to perform a name replacement on the
debug file of the first loaded module.

Note that this renaming of the debug file is transient—it's only used for lookups and
reverted before we return the symbolication response.