feat: Add support for Symstore indexes (#1663)
This adds the ability for debug file sources to indicate that they have an
index via the `has_index` flag. Currently, only one index type (Symstore)
is supported, and it's only enabled for sources using the Symstore layout.

If a source has an index, it will be computed before fetching files from the source.
Failure to fetch the index results in no requests being made to the source. For that
reason, failed index timeouts are retried after a short time (10min by default).