fix(index): Improve Symstore last ID logic (#1686)
The last ID fetching logic was set up so that the result (good or bad) of fetching the last ID from the index would be cached for 24h. This had some problems:

* The 24h period was set up with _successful_ downloads in mind. If the fetching fails, we likely want to retry much sooner than that.
* A download failure would return an error for the whole index, even if there is a perfectly good index still available for the previous last ID.

Really, what we want is something like this:
* If downloading the last ID succeeded, cache it for a day.
* If downloading the last ID failed and there is no previous good value, cache the error for 10min.
* If downloading the last ID failed and there is a previous good value, reuse it for another 24 hours.

This means that as soon as a download of the last ID succeeds once, it will be reused in perpetuity until a new download succeeds.

This PR implements this logic by way of the `moka` `and_compute_with` API.