feat(symsorter): Save separate metadata for each file (#1676)
In cases where multiple input files are sorted into the same folder,
e.g. app.exe and its corresponding app.pdb, the meta file only contains
information of the last file written.  Generate separate `.meta` files
for each sorted file so that this information is not lost.

This would break integrations that assumes the existence of the `meta`
file, but this behaviour change should be obvious and trivial to adapt
to.