feat: allow sending `bearer_token` in request instead of providing credentials (#1697)
This PR extends the request format to allow sending a bearer_token directly instead of providing private_key and client_email so that symbolicator fetches the token itself.

The supplied token will not be cached or re-used and needs to be supplied on every request when using this as preferred authorization.