ref: Unify download timeout configuration (#1709)
Previously we had individual download timeout settings in the config
and then a struct bundling them in the `utils` module. The default
values were inconsistent between them (meaning the ones on the struct
didn't matter) and this was just unnecessary duplication in general.

This moves the `DownloadTimeouts` struct into the config. No changes to
config parsing.