feat(proguard): Update proguard cache format
This updates the `proguard` dependency to 5.6.1 to get access to
"synthesized" flags in the cache format
(see https://github.com/getsentry/rust-proguard/pull/52).

Since the cache format has changed, this requires an _incompatible_
bump of the proguard cache version.

It also adds a static version check for the cache version exposed by the
`proguard` crate in analogy to symcaches and cficaches.