fix(ci): Align image/assemble and assemble-ar jobs
This changes the `assemble` job to run on `ubuntu-latest` (same as
`assemble-ar`) and the `assemble-ar` job to also push the image as
`:nightly` (same as `assemble`).

The latter part in particular is important because the `devservices`
config expects the current image to be found at
`us-central1-docker.pkg.dev/sentryio/symbolicator/image:nightly`, and
that image was last pushed on 2025-01-17.