Add support for fetching minidumps directly from objectstore
Instead of receiving attachments via a multipart upload, this will instead fetch the minidump via the `objectstore-client`. For that, it currently still transfers the necessary metadata via yet another multipart field.

Eventually, the minidump endpoint should transition over to a JSON body instead of multipart, in the distant future once this is fully rolled out.