build: Update dependencies (#1781)
* The `axum` update requires a few extra `Sync` bounds and a change in
how captures in routes are written.
* The AWS SDK update requires us to manually initialize a
`CryptoProvider` (from `rustls`) at the start of every process. Up until
now, we depended on `rustls` with just the `ring` feature enabled, which
let it initialize a `CryptoProvider` automatically. But the AWS SDK now
pulls in `rustls` with two different features enabled (`ring` and
`aws-lc-rs`), so it can no longer automatically pick a provider.