Bump even more dependencies (#1786)
This manually bumps more dependencies that were missed last time, as
well as running a proper `cargo update`, which cleans up duplicated
crate versions, etc.

Another change here is that the `reqwest` fork was updated to the branch
matching https://github.com/getsentry/reqwest/pull/7, which is a refresh
of our previous patch-set.

This is quite a large change, so should be landed with caution.