Symsorter always prints "Created 0 source bundles" (#1787)
... the original implementation would collect partial sums by
multiplying them, instead of adding them together.

This resulted in a series of multiplications by 0 -- and therefore, it
would always print "Created 0 source bundles" at the end.

Here's a fix for that. This does not affect the functionality of the
tool in any other way; just the stdout message at the end.