Add a unified `symbolicate-v2` endpoint
The aim of this endpoint is to eventually supercede all other endpoints,
unifying common request fields in a top-level body, and then matching on the actual symbolication request.

As a first usecase, this now supports new `objectstore`-based requests for minidump and apple crashreport processing.
This can thus replace the existing multipart based endpoints, and takes the `storage_url` as part of the JSON body.

It then downloads the given file, and symbolicates as usual.