fix(download): Add dartsymbolmap to sentry file types (#1803)
Symbolicator fails to parse a list of files from the Sentry symbol
source when they contain a dartsymbolmap, which was added in this PR:
https://github.com/getsentry/sentry/pull/97076.

Symbolicator is not supposed to do anything with this symbol map, it is
used in sentry preprocessing. Adding the enum variant should fix the
parsing error.