Add a unified `symbolicate-any` endpoint (#1795)
The aim of this endpoint is to eventually supercede all other endpoints,
unifying common request fields in a top-level body, and then matching on
the actual symbolication request.

As a first usecase, this now supports new `objectstore`-based requests
for minidump and apple crashreport processing. This can thus replace the
existing multipart based endpoints, and takes the `storage_url` as part
of the JSON body.

It then downloads the given file, and symbolicates as usual.

---

This is an alternative to
https://github.com/getsentry/symbolicator/pull/1780, and avoids pulling
in our `objectstore-client`.
It also creates a new endpoint instead of reusing the existing
multipart-based ones.

Ref FS-103