ref(proguard): Clarify stacktrace order
Sentry orders stacktraces from outermost frame to innermost.
Symbolicator does it the other way around, from innermost to outermost.
For native events, we reverse stacktraces in Sentry before sending them
to Symbolicator and then reverse the result again. This means that
Symbolicator can just assume the order it wants and never has to reverse
anything.

I neglected to take this into account when porting proguard processing
to Symbolicator, so Sentry sends JVM stacktraces in thge order it wants.
Instead, we previously reversed the output of the `map_full_frame`
function, which achieved the same result.

This PR attempts to clean up the situation by moving the reversing logic
as far outward as possible. JVM stacktraces are now reversed (from
"Sentry order" to "Symbolicator order" in the `symbolicate-jvm`) endpoint
and reversed again when the response is returned. This means that in the
"interior" of Symbolicator the order is always the same.

The PR also changes a number of tests: it introduces snapshots for some
that were previously asserting stuff and reverses some to get the
"expected" order.

In the future we might switch to reversing stacktraces in Sentry, like
it already works for native.