feat: Make request body size configurable
This adds configuration options `crash_file_body_max_bytes` and
`symbolicate_body_max_bytes` that determine the maximum body sizes for
requests to the `/minidump`, `/applecrashreport`, `/symbolicate`,
`/symbolicate-any`, `/symbolicate-js`, and `/symbolicate-jvm` endpoints.

The default values for these options are 250MiB and 10MiB, respectively.
This is prompted by the fact that we now accept 200MiB minidumps in
Sentry.