feat: Make order of stack frames configurable (#1823)
An expansion/more principled version of #1820.

This adds a new enum `FrameOrder` that can be sent with symbolication
requests. The possible values are `"callee_first"` and `"caller_first"`.
`"caller_first"` is the order in which Sentry stores stacktraces in
events, i.e. the innermost frame is the last.

This option determines how Symbolicator will interpret the stacktraces
it receives (and it replies in the same order). Currently, what Sentry
sends and Symbolicator assumes is, to put not too fine a point on it, an
inconsistent and undocumented mess.

This PR significantly improves the situation by making the order of
frames explicit whenever a symbolication request struct is constructed
within Symbolicator. Down the road we can use this to get rid of the
reversing of stacktraces in Sentry.

There is no global default for `FrameOrder`. Instead, each endpoint
specifies the default individually in such a way that the current behavior is preserved.

Fixes #1822. Fixes SYMBOLI-38.