feat(proguard): Support rewriteFrame annotation (#1845)
Attempts to reuse `remap_frame` from ProguardCache to delegate as much
logic to it as possible and keep it lean on the caller side.

Additionally, had to extend the `JvmStacktrace` interface, because we
now need the `JvmException` associated with the stacktrace to correctly
apply rewrite rules.

Depends on https://github.com/getsentry/rust-proguard/pull/67